pyproject.toml
[project]
name = "epicgamesranking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachier>=4.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
